refactor: i18n
no circular dependencies
